backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
redis[hiredis]==5.2.1
python-jose[cryptography]==3.3.0
aiogram==3.17.0
qrcode[pil]==8.0
structlog==24.4.0
python-multipart==0.0.20